[libwspcodec] Allow empty quoted subject. JB#57441
Some software encodes empty MMS subject as 2 bytes: 0x7f 0x00.
That's Quote and End-of-string characters according to WSP spec:

  Quote = <Octet 127>
  End-of-string = <Octet 0>

However the same section of WSP spec says this:

  Text-string = [Quote] *TEXT End-of-string

  If the first character in the TEXT is in the range of 128-255,
  a Quote character must precede it. Otherwise the Quote character
  must be omitted. The Quote is not part of the contents.

Since the subject is empty, the first character in the TEXT is not
in the range of 128-255 (since there's no first character), therefore
the Quote character must be omitted. But it's there. Such encoding is
incorrect.

But let's allow it anyway.